Add files via upload
The project answers three questions that are derived from the GSS data. They are if or not religion affects the confidence of people in the legislature, how different has finding equally good jobs been in 2002 and 2012 and does being above or below 20 years old affect the consideration of committing suicide under different circumstances.